% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportData.r
\name{export,ExportData-method}
\alias{export,ExportData-method}
\title{Title}
\usage{
\S4method{export}{ExportData}(self)
}
\arguments{
\item{self}{ExportData.une classe avec 2 attributs monjdd et filename}
}
\value{
un commentaire et un fichier au bon endroit
}
\description{
Title
}
\examples{
monjdd_ex<-data.frame(Pays = c("Allemagne", "Japon", "Norvege"),Capitale = c("Berlin", "Tokyo", "Oslo"));monoutput<- new("ExportData", monjdd="monjdd_ex", filename="montest_exemple.csv");export(monoutput)
}
