% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.r
\name{ImportData}
\alias{ImportData}
\title{Importe mon jeu de donnees}
\usage{
ImportData(chemin, colonnes = NULL, numeriques = NULL, sep = ";", dec = ",")
}
\arguments{
\item{chemin}{specifier le chemin ou se trouve le fichier}

\item{colonnes}{detail des colonnes à importer(facultatif)}

\item{numeriques}{numéro de colonnes (facultatif)}

\item{sep}{separateur des colonnes par defaut ";"}

\item{dec}{decimales par defaut ","}
}
\value{
un jeu de donnees lisible en r
}
\description{
Importe mon jeu de donnees
}
\examples{
ImportData("./data/output/montest_export.csv")
}
